+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # CHANGELOG
 
 ## 0.1.1dev
-* [Feature] Adds `pkgmt check-project`
+* [Feature] Adds `pkgmt check`
 * [Feature] Adds `pkgmt hook`
+
 ## 0.1.0 (2023-01-06)
 * Completing version string when missing components. (e.g., `1.0` -> `1.0.0`) ([#18](https://github.com/ploomber/pkgmt/issues/18))